Updating lots of inline documentation, tiny refactor
I'm working on making sure the doc comments are all present. I also
moved the parsing files out of subdirectories to keep the file structure
more consistent.
