+++ src/modules/glyphs/line/line-defaults.ts
@@ -3,14 +3,35 @@ import {Annotation} from "../../../annotations/annotation";
 import {ZoomBehavior} from "../../zoom/zoom-behavior";
 import {Chart} from "../../../charts/chart";
 
+/**
+ * @hidden
+ */
 export const horizontalLineX1Fn = <A extends Annotation>(a: A) => a.getX();
+/**
+ * @hidden
+ */
 export const horizontalLineX2Fn = <A extends Annotation>(a: A) => a.getX() + a.getW();
+/**
+ * @hidden
+ */
 export const horizontalLineYFn = <A extends Annotation, C extends Chart<any>>(a: A, c: C) => a.y * c.binHeight + c.binHeight / 2;
 
+/**
+ * @hidden
+ */
 export const verticalLineXFn = <A extends Annotation>(a: A) => (a.getX() + a.getW() / 2);
+/**
+ * @hidden
+ */
 export const verticalLineY1Fn = <A extends Annotation, C extends Chart<any>>(a: A, c: C) => a.y * c.binHeight;
+/**
+ * @hidden
+ */
 export const verticalLineY2Fn = <A extends Annotation, C extends Chart<any>>(a: A, c: C) => (a.y + 1) * c.binHeight;
 
+/**
+ * @hidden
+ */
 export class LineZoomBehavior<A extends Annotation, C extends Chart<any>> implements ZoomBehavior<C, d3.Selection<SVGLineElement, A, HTMLElement, any>> {
     // the default ZoomBehavior for a vertical line drawn with this module
     // it moves and stretches the line horizontally
