+++ src/modules/glyphs/plots/axis-defaults.ts
@@ -3,8 +3,14 @@ import {ZoomBehavior} from "../../zoom/zoom-behavior";
 import * as d3 from "d3";
 import {Annotation} from "../../../annotations/annotation";
 
+/**
+ * @hidden
+ */
 export const axisXFn = <A extends Annotation>(a: A) => a.getX();
 
+/**
+ * @hidden
+ */
 export class VerticalAxisZoomBehavior<A extends Annotation, C extends Chart<any>> implements ZoomBehavior<C, d3.Selection<SVGGElement, A, HTMLElement, any>> {
     id: string = 'default-vertical-axis-zoom-behavior';
     selector: string;
