+++ src/modules/glyphs/sequence/sequence-defaults.ts
@@ -3,10 +3,22 @@ import {Chart} from "../../../charts/chart";
 import {ZoomBehavior} from "../../zoom/zoom-behavior";
 import {SequenceAnnotation, CharacterDatum} from "../../../annotations/sequence-annotation";
 
+/**
+ * @hidden
+ */
 export const sequenceXFn = <A extends SequenceAnnotation, C extends Chart<any>>(a: A) => a.getX();
+/**
+ * @hidden
+ */
 export const characterXFn = <D extends CharacterDatum, C extends Chart<any>>(d: D) => d.x;
+/**
+ * @hidden
+ */
 export const sequenceYFn = <A extends SequenceAnnotation, C extends Chart<any>>(a: A, c: C) => a.y * c.binHeight;
 
+/**
+ * @hidden
+ */
 export class SequenceZoomBehavior<A extends SequenceAnnotation, C extends Chart<any>> implements ZoomBehavior<C, d3.Selection<SVGGElement, A, HTMLElement, any>> {
     selector: string;
     id = 'default-sequence-zoom-behavior';
@@ -33,6 +45,9 @@ export class SequenceZoomBehavior<A extends SequenceAnnotation, C extends Chart<
     }
 }
 
+/**
+ * @hidden
+ */
 export class SequenceCharacterZoomBehavior<D extends CharacterDatum, C extends Chart<any>> implements ZoomBehavior<C, d3.Selection<SVGTextElement, D, HTMLElement, any>> {
     selector: string;
     id = 'default-sequence-character-zoom-behavior';
