+++ src/modules/id-map/id-map.ts
@@ -1,11 +1,18 @@
 import * as d3 from 'd3';
 import { Annotation } from '../../annotations';
 
-// this module provides some convenience maps that give
-// access to selections and annotations via id
+/**
+ * @hidden
+ */
 export const idAnnotationMap: Map<string, Annotation> = new Map();
+/**
+ * @hidden
+ */
 export const idSelectionMap: Map<string, d3.Selection<any, any, any, any>> = new Map();
 
+/**
+ * @hidden
+ */
 export function getAllIds(): string[] {
     return Array.from(idAnnotationMap.keys());
 }
