+++ src/modules/layout/greedy-graph-layout.ts
@@ -1,6 +1,9 @@
 import {Annotation} from "../../annotations/annotation";
 import {AnnotationGraph} from "./annotation-graph";
 
+/**
+ * @hidden
+ */
 const DEFAULT_VERTEX_SORT = (verts: string[], graph: AnnotationGraph<Annotation>) => {
     // the default vertex sort function; it just sorts by annotation width
     verts.sort((v1: string, v2: string) => {
