+++ src/modules/layout/heuristic-graph-layout.ts
@@ -78,6 +78,9 @@ export function heuristicGraphLayout(ann: Annotation[], nIters: number = 100, to
     return (nextColor);
 }
 
+/**
+ * @hidden
+ */
 function shuffle(a: string[]) {
     for (let i = a.length - 1; i > 0; i--) {
         const j = Math.floor(Math.random() * (i + 1));
