+++ src/modules/parse.ts
@@ -1,2 +1,4 @@
-export * from './parse/gff3'
-export * from './parse/bed'
+export * from './parse/gff3-parse'
+export * from './parse/gff3-record'
+export * from './parse/bed-parse'
+export * from './parse/bed-record'
