--- src/modules/parse/bed.ts
@@ -1,2 +1,0 @@
-export * from './bed/bed-parse'
-export * from './bed/bed-record'
