+++ src/modules/parse/gff3-record.ts
@@ -1,3 +1,6 @@
+/**
+ * An interface that describes the SODA representation of a GFF3 record.
+ */
 export interface Gff3Record {
     seq_id: string,
     source: string,
