--- src/modules/parse/gff3.ts
@@ -1,2 +1,0 @@
-export * from './gff3/gff3-parse'
-export * from './gff3/gff3-record'
